fixed action saving
