+++ src/lib/types.ts
@@ -1,0 +1,7 @@
+type Action = {
+    id: number;
+    name: string;
+    score: number;
+    category: string;
+    image: string;
+};
